Move Callum's code to a dedicate script | Created a keyword search algorithm which searches for keywords and choose relevant scripts to use in order to respond
In order to decluter the chatbot_script.py, I move Callum's code to a dedicated script in which I also added some functions that could generate a basic reply to the end user based off of the context.

This context was based off of an algorthim which I made which would scan the user's message for keywords and call scripts for relevant replies.

The bot also chains messages. So if it is thanked and then told goodbye, it will respond to the thanks and send a goodbye message.

- Added Appreciation Response (Bot responds to the user relevantly if thanked)

- Added Greeting Response (Bot responds to the user relevantly if greeted)

- Added Farewell Response (Bot responds to the user if told farewell, bye or goodbye.)

TL;DR - The bot can give very basic responses to user input. Moved relevant code to relevant scripts.